spawn function cleanup + new effect
2 New effects have been added:
- Random Crit, Chance for a Grenade launcher to deal 3x damage,
- Cursed, Chance to get inverted controls when using adrenaline

Cleaned up how entities are spawned e.g jumpscare
Instead of overwriting command flags to make spawning commands non cheat temporarly and executing them on the client
we use a native from Left 4 dhooks direct to achieve the same. this also allows for keeping track of the spawned entities and not accidentally kill a director spawned for example.